Added the concept of subject (rat) to the experiment framwork and specifyer xml file.
Added plotting capabilities to the framework. One plotting script is copied to the experiment loggin folder, which generates plots for all trials and individuals. TODO: execute it automatically at the end of the experiment.

Some file as maze and configuration are copied to the experiment log folder for plotting/traceability purposes.